refactor: arrow common component 폴더 이동 및 구조 변경
