+++ src/components/Slider/Slider.components/_Stepper.tsx
