+++ src/components/Slider/Slider.styles.ts
@@ -1,0 +1,22 @@
+import styled from "@emotion/styled";
+
+export const ArrowContainer = styled.div`
+  position: absolute;
+  margin: auto;
+  top: 0;
+
+  display: flex;
+  width: 32px;
+  height: 32px;
+  justify-content: center;
+  align-items: center;
+  border-radius: 32px;
+  overflow: visible;
+
+  border: 1px solid gray;
+  background-color: #fff;
+  box-shadow: 0px 1px 5px -3px rgba(0, 0, 0, 0.04);
+
+  cursor: pointer;
+  z-index: 1;
+`;
